Move QT Applet to new home under arrows
