+++ arrows/qt/applets/CMakeLists.txt
@@ -1,0 +1,4 @@
+
+if( KWIVER_ENABLE_PROCESSES )
+  add_subdirectory( pipeline_viewer )
+endif()
