+++ arrows/qt/applets/pipeline_viewer/ImageView.cxx
