+++ arrows/qt/applets/pipeline_viewer/MainWindow.cxx
