+++ arrows/qt/applets/pipeline_viewer/PipelineWorker.cxx
