+++ arrows/qt/applets/pipeline_viewer/pipeline_viewer.h
@@ -29,52 +29,29 @@
  * DAMAGE.
  */
 
-#include "pipeline_viewer.h"
+#ifndef KWIVER_TOOL_PIPELINE_VIEWER_H
+#define KWIVER_TOOL_PIPELINE_VIEWER_H
 
-#include "MainWindow.h"
+#include <vital/applets/kwiver_applet.h>
 
-#include <vital/range/iota.h>
+#include <string>
+#include <vector>
 
-#include <QApplication>
+namespace kwiver {
+namespace tools {
 
-// ----------------------------------------------------------------------------
-void
-kwiver::tools::pipeline_viewer
-::add_command_options()
+class pipeline_viewer : public kwiver_applet
 {
-  skip_parse_args();
-}
+public:
+  PLUGIN_INFO( "pipe-gui",
+               "Run pipelines in a simple GUI.\n\n"
+               "This program provides a simple Qt-based front-end "
+               "for executing pipelines and viewing images produced by the same." );
 
-// ----------------------------------------------------------------------------
-int
-kwiver::tools::pipeline_viewer
-::run()
-{
-  std::vector< std::string > const& args = applet_args();
-  // Create mutable arguments (needed by QApplication)
-  auto argc = static_cast< int >( args.size() );
-  auto argv = std::unique_ptr< char*[] >{ new char*[ argc ] };
+  virtual int run() override;
+};
 
-  auto margs = std::vector< std::unique_ptr< char[] > >{};
-  for ( auto const i : vital::range::iota( args.size() ) )
-  {
-    auto const& arg = args[ i ];
-    auto const l = arg.size() + 1;
+} // namespace tools
+} // namespace kwiver
 
-    auto marg = std::unique_ptr< char[] >{ new char[ l ] };
-    memcpy( marg.get(), arg.c_str(), l );
-
-    argv[ i ] = marg.get();
-    margs.push_back( std::move( marg ) );
-  }
-
-  // Create QApplication
-  QApplication app{ argc, argv.get() };
-
-  // Create and show main window
-  kwiver::tools::MainWindow window;
-  window.show();
-
-  // Execute event loop
-  return app.exec();
-}
+#endif
